首页获取不同类型的文章简介

+++ src/main/java/com/zhixin/vo/request/RequestModuleSearchVo.java
@@ -9,9 +9,9 @@ import lombok.Data;
 @Data
 public class RequestModuleSearchVo {
 
-    private Integer typeId;
+    private Long typeId;
 
-    private Integer subTypeId;
+    private Long subTypeId;
 
     private Integer pageStart;
